backend/auth_service/requirements.txt
fastapi
aioredis
pydantic-settings
kink
uvicorn[standard]
databases[asyncpg]